Create sha256.js
Credit to Ast (@Ast#9365 on discord) for this. Most of my scripts, as of now, could not have been done without this one here. Ast, if you want me to take this down, let me know and I will do so. I am only putting this up so that other people can use my scripts based on it. My suggestion to you, ast, is to put up the script publicly yourself, so I can change all my references from here to yours.